requirements.txt
requests
deepdiff
jinja2
python-dotenv
flask